Wishing Happy Birthday in Python
The above two python files HappyBirthday1.py and happybday.py wishes the happy bday messages to you in two different manners. 
while the first one uses Turtle graphics module to print the interactive bday cake with a happy birthday message while the later uses the for loops to print a cake like pattern where no of layers to be displayed are given as input by the user. 
Here is the link of an online Python Interpreter where both the programs can be run.
https://repl.it/@rajpratyush/WateryHappyHashmaps#main.py  (HappyBirthday1.py)
https://repl.it/@rajpratyush/UnderstatedLiveBetaversion#main.py   ( happybday.py)